Teach elmer to chat with the Snowflake Cortex Analyst.
This commit introduces `chat_cortex()`, a new provider that allows the
user to chat with the Snowflake Cortex Analyst [0].

The Analyst has a bespoke -- and underdocumented -- API, so a fair
amount of this is custom handling determined by trial and error. The
major distinction of Cortex Analyst is that it returns a structured
response that can contain text, SQL, and "suggestions" -- rather than
simply a block of text. elmer is somewhat equipped to handle this
already, though I had to make changes in a few places to accomodate it.

Under the hood, the Analyst is powered by Llama and Mistral models, but
the caller doesn't specify the model at all. Nor do they provide a
system prompt -- instead, the caller points to something that Snowflake
calls a "semantic model" [1], which is a YAML file with a known
specification.

The Analyst also doesn't support multi-turn, though I expect it will in
the future (and the error message you get when trying indicates as
much). There is also no support for tool calling.

I've implemented (and verified) both authentication methods that the API
supports, namely Snowflake OAuth and what they term "key-pair"
authentication, which uses JWTs signed by an RSA key. Key-pair
authentication can be especially useful because it should allow us to
run tests against Cortex in CI.

Unit tests are included. Integration tests also run if there are ambient
Snowflake credentials available.

[0]: https://docs.snowflake.com/en/user-guide/snowflake-cortex/cortex-analyst
[1]: https://docs.snowflake.com/en/user-guide/snowflake-cortex/cortex-analyst/semantic-model-spec

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>