Give more control over echoing (#78)
* Give more control over echoing

* Reimplement line prefixing when catting

Before this commit, the cat_word_wrap_impl had some bugs
in edge cases. I tried to fix them surgically but it
became impossible for me to understand the interactions
between the prefixing and the word wrapping. This commit
divides the functionality of cat_word_wrap into two
separate "sink decorators", with cat() at the center,
a line prefixing decorator around it, and a word wrap
decorator around that.

It's a fair amount more lines of code this way, but the
individual decorators are easier to understand and get
the edge cases exactly right.

* Add news bullet

---------

Co-authored-by: Joe Cheng <joe@posit.co>