Outsource Snowflake authentication for Cortex to `snowflakeauth`.
This commit drops all of the Snowflake-specific authentication logic in
favour of using the new `snowflakeauth` package. This has a big
advantage in that it works with Snowflake connections and credentials
created by the Snowflake Python packages and their CLI.

However, this is a breaking change in that (1) some environment
variables that previously furnished credentials no longer do; and (2)
the signature of `chat_cortex()` has changed. The reason to drop the
environment variable support is that these specific environment
variables are not used by Snowflake's Python tooling -- so I think this
is the right choice.

Existing unit tests do cover regressions in the Authorization headers we
pass to Snowflake, though I did have to modify them a bit to account for
the API changes.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>