Add support for chatting with models hosted on Databricks.
This commit adds support for chatbots that work with models hosted on
Databricks [0].

According to their documentation, Databricks models can be queried by
the OpenAI Python SDK, so in theory they would be compatible with our
OpenAI provider. But in practice there are enough differences that a new
provider is necessary.

There are other limitations, too: structured outputs are in public
preview, and there is no image support.

Finally, I've found that tool calling support (also in public preview)
is very limited, too limited for `elmer` to work properly at all:

    > chat <- chat_databricks()
    > chat$register_tool(tool(function() "2024-01-01", "Gets the current date"))
    Using model = "databricks-dbrx-instruct".
    > chat$chat("What's the current date in YMD format?")
    Error in `req_perform_connection()` at elmer/R/httr2.R:36:3:
    ! HTTP 400 Bad Request.
    • Bad request: Model does not support continuing the chat past the first tool call
    Show Traceback
    > chat
    <Chat turns=2 tokens=726/11>
    ── user ────────────────────────────────────────────────────────────────────────────────────────
    What's the current date in YMD format?
    ── assistant ───────────────────────────────────────────────────────────────────────────────────
    [tool request (call_93d2e2fe-1a2e-4ef4-80df-4feaed245569)]: tool_001()

Databricks has announced that this limition will be lifted once tool
calling is out of public preview [1], so I suppose we'll have to cross
our fingers that this will start working at that point.

Unit tests and docs are included.

[0]: https://docs.databricks.com/en/machine-learning/model-serving/score-foundation-models.html
[1]: https://docs.databricks.com/en/machine-learning/model-serving/function-calling.html#limitations

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>